Tried getting scissoring to work.

+++ src/main.rs
@@ -73,7 +73,7 @@ fn main() {
                         }
                     },
                     WindowEvent::CursorMoved { position, .. } => {
-                        let size = app_state.screen.get_tile_size();
+                        let size = app_state.screen.get_tile_size() as f32;
                         let mut delta = vector((position.x - last_pos.x) as f32, (position.y - last_pos.y) as f32);
                         let zoom_x = (app_state.screen.width as f32) / size / 2f32.powf(app_state.zoom) / size / 2.0 / 1.3;
                         let zoom_y = (app_state.screen.height as f32) / size / 2f32.powf(app_state.zoom) / size / 2.0 / 1.3;
